extracted collection prior model
