+++ test/mapper/test_prior_model.py
@@ -1,4 +1,5 @@
 import autofit.mapper.prior_model.abstract_prior_model
+import autofit.mapper.prior_model.collection_prior_model
 from autofit import mapper as m
 from test.mock import SimpleClass, ComplexClass, ListClass, Distance, \
     DistanceClass, PositionClass, Galaxy, Tracer
@@ -214,7 +215,7 @@ class TestCase(object):
 
 class TestCollectionPriorModel(object):
     def test_keyword_arguments(self):
-        prior_model = m.CollectionPriorModel(
+        prior_model = autofit.mapper.prior_model.collection_prior_model.CollectionPriorModel(
             one=SimpleClass,
             two=SimpleClass(1, 2)
         )
@@ -236,7 +237,7 @@ class TestCollectionPriorModel(object):
         assert instance.two.two == 2
 
     def test_mix_instances_in_list_prior_model(self):
-        prior_model = m.CollectionPriorModel([SimpleClass, SimpleClass(1, 2)])
+        prior_model = autofit.mapper.prior_model.collection_prior_model.CollectionPriorModel([SimpleClass, SimpleClass(1, 2)])
 
         assert len(prior_model.prior_models) == 1
         assert prior_model.prior_count == 2
@@ -256,25 +257,25 @@ class TestCollectionPriorModel(object):
         assert len(prior_model.prior_class_dict) == 2
 
     def test_list_in_list_prior_model(self):
-        prior_model = m.CollectionPriorModel([[SimpleClass]])
+        prior_model = autofit.mapper.prior_model.collection_prior_model.CollectionPriorModel([[SimpleClass]])
 
         assert len(prior_model.prior_models) == 1
         assert prior_model.prior_count == 2
 
     def test_list_prior_model_with_dictionary(self):
-        prior_model = m.CollectionPriorModel({"simple": SimpleClass})
+        prior_model = autofit.mapper.prior_model.collection_prior_model.CollectionPriorModel({"simple": SimpleClass})
 
         assert isinstance(prior_model.simple, m.PriorModel)
 
     def test_labels(self):
         mapper = m.ModelMapper()
 
-        mapper.my_list = m.CollectionPriorModel({"simple": SimpleClass})
+        mapper.my_list = autofit.mapper.prior_model.collection_prior_model.CollectionPriorModel({"simple": SimpleClass})
 
         assert mapper.info.split("\n")[4].startswith("my_list_simple_one")
 
     def test_override_with_constant(self):
-        prior_model = m.CollectionPriorModel({"simple": SimpleClass})
+        prior_model = autofit.mapper.prior_model.collection_prior_model.CollectionPriorModel({"simple": SimpleClass})
 
         simple_instance = SimpleClass(1, 2)
